requirements.txt
Django==1.7.2
Pillow==2.7.0
sorl-thumbnail==12.2
wsgiref==0.1.2
